+++ src/Surface.ts
@@ -25,6 +25,28 @@ namespace g {
 		isDynamic: boolean;
 
 		/**
+		 * 本Surfaceの画像がスケール変更可能かを示す値。真の時、スケール変更可能。
+		 * この値は参照のみに利用され、変更してはならない。
+		 */
+		hasVariableResolution: boolean;
+
+		/**
+		 * 本Surfaceのx方向のスケール値
+		 */
+		xScale: number;
+
+		/**
+		 * 本Surfaceのy方向のスケール値
+		 */
+		yScale: number;
+
+		/**
+		 * 本Surfaceの再描画イベント
+		 * hasVariableResolutionが偽の時undefined
+		 */
+		contentReset: Trigger<void>;
+
+		/**
 		 * アニメーション再生開始イベント。
 		 * isDynamicが偽の時undefined。
 		 */
@@ -54,19 +76,29 @@ namespace g {
 		 * @param width 描画領域の幅（整数値でなければならない）
 		 * @param height 描画領域の高さ（整数値でなければならない）
 		 * @param drawable 描画可能な実体。省略された場合、 `undefined`
-		 * @param isDynamic drawableが動画であることを示す値。動画である時、真を渡さなくてはならない。省略された場合、偽。
+		 * @param optionalFlag 本surfaceに関するフラグをまとめたもの。詳細は以下の通り。
+		 * 1ビット目：drawableが動画であることを示す値、2ビット目：本Surfaceの画像がスケール変更可能かを示す値
+		 * また、互換性を保つためbooleanも許可している。
 		 */
-		constructor(width: number, height: number, drawable?: any, isDynamic: boolean = false) {
+		constructor(width: number, height: number, drawable?: any, optionalFlag: number|boolean = false) {
 			if (width % 1 !== 0 || height % 1 !== 0) {
 				throw ExceptionFactory.createAssertionError("Surface#constructor: width and height must be integers");
 			}
 
 			this.width = width;
 			this.height = height;
+			this.xScale = 1;
+			this.yScale = 1;
 			if (drawable)
 				this._drawable = drawable;
-			// this._destroyedは破棄時に一度だけ代入する特殊なフィールドなため、コンストラクタで初期値を代入しない
-			this.isDynamic = isDynamic;
+			const normalizedOptionalFlag = Number(optionalFlag) & parseInt("11", 2);
+			this.isDynamic = Boolean(normalizedOptionalFlag & parseInt("1", 2));
+			this.hasVariableResolution = Boolean(normalizedOptionalFlag & parseInt("10", 2));
+			if (this.hasVariableResolution) {
+				this.contentReset = new Trigger<void>();
+			} else {
+				this.contentReset = undefined;
+			}
 			if (this.isDynamic) {
 				this.animatingStarted = new Trigger<void>();
 				this.animatingStopped = new Trigger<void>();
@@ -74,6 +106,7 @@ namespace g {
 				this.animatingStarted = undefined;
 				this.animatingStopped = undefined;
 			}
+			// this._destroyedは破棄時に一度だけ代入する特殊なフィールドなため、コンストラクタで初期値を代入しない
 		}
 
 		/**